refactor: 💡 렌딩페이지의 방개설 버튼 연결
